+++ src/main/java/io/spring/training/boot/trainingplan/web/dto/AthleteResponseDto.java
@@ -2,9 +2,9 @@ package io.spring.training.boot.trainingplan.web.dto;
 
 public record AthleteResponseDto(
         long id,
-        String socialIdentifier,
+        String cpf,
         String firstName,
         String lastName,
         int age,
-        double weight
+        double weightKg
 ) {}