+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## \[0.0.3] - 2025-08-21
+
+### Added
+- Created test environment with the following libraries
+  - Mockito
+  - JUnit5
+  - MockMvc
+  - Mockito.JUnit
+  - AssertJ
+  - Testcontainers
+  - Testcontainers.JUnit
+  - Testcontainers.postgres
+- Created 3 tests for `AthleteController` using web slice
+- Created 2 tests for `AthleteMapper`
+
+### Fix
+- `Athlete` Using JPA documentation, adjusted Lombok annotations for not encountering bugs or delay in performance
+- `AthleteMapper` mapped socialIdentifier to CPF in the conversion entity to dto
+
+### Changed
+- `AthleteResponseDto` changed socialIdentifier field to CPF
+
 ## \[0.0.2] - 2025-08-13
 
 ### Added